[Rename] card 모듈 internal 정리
