+++ src/main/kotlin/com/whatever/caro/card/internal/deck/DeckPresetRepository.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.card.internal
+package com.whatever.caro.card.internal.deck
 
 import org.springframework.data.jpa.repository.JpaRepository
