+++ src/main/kotlin/com/whatever/caro/card/internal/notetype/CardTemplate.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.card.internal
+package com.whatever.caro.card.internal.notetype
 
 import com.whatever.caro.common.entity.BaseTimeEntity
 import jakarta.persistence.Column
