+++ src/main/kotlin/com/whatever/caro/card/internal/notetype/NoteTypeRepository.kt
@@ -1,4 +1,4 @@
-package com.whatever.caro.card.internal
+package com.whatever.caro.card.internal.notetype
 
 import org.springframework.data.jpa.repository.JpaRepository
